added ability to restore users

+++ app.ts
@@ -5,9 +5,8 @@ import * as User from "./User.js";
 import chalk from "chalk";
 import { chownSync } from "fs";
 
-const version = '0.4.0'
+const version = "0.5.0";
 const commands: cli.CLICommand[] = [
-
     {
         text: `help`,
         callback: (args) => {
@@ -29,7 +28,7 @@ const commands: cli.CLICommand[] = [
     {
         text: `read`,
         callback: (args) => {
-            read(args)
+            read(args);
         },
     },
     {
@@ -47,9 +46,17 @@ const commands: cli.CLICommand[] = [
         },
     },
     {
+        text: `restore`,
+        callback: (args) => {
+            restore(args);
+            list();
+        },
+    },
+    {
         text: `update`,
         callback: (args) => {
             update(args);
+            list();
         },
     },
     {
@@ -61,23 +68,33 @@ const commands: cli.CLICommand[] = [
     {
         text: `version`,
         callback: (args) => {
-            if (args[0] === '🥚') easter.egg(args);
+            if (args[0] === "🥚") easter.egg(args);
             cli.version(version);
         },
-    }
+    },
 ];
 
 function update(args: cli.ArgumentsList) {
-
     const searchIndex = Number(args[0]);
 
-    if (!searchIndex) cli.log(chalk.green, `ℹ️ Please update using user index.\n\n${chalk.bold('update [user_id] [field] [value]')}\n`)
+    if (!searchIndex)
+        cli.log(
+            chalk.green,
+            `ℹ️ Please update using user index.\n\n${chalk.bold(
+                "update [user_id] [field] [value]"
+            )}\n`
+        );
 
     const user = User.getUsers()[searchIndex];
 
     if (!user) {
-        cli.log(chalk.red, chalk.bold(`⚠️  Couldn't find user with id ${chalk.underline(searchIndex)}\n\n`))
-        exit()
+        cli.log(
+            chalk.red,
+            chalk.bold(
+                `⚠️  Couldn't find user with id ${chalk.underline(searchIndex)}\n`
+            )
+        );
+        exit();
     }
 
     const key = User.isEditableUserKey(args[1]);
@@ -85,13 +102,24 @@ function update(args: cli.ArgumentsList) {
 
     if (key) {
         User.update(user, key, value);
+        User.saveUsers();
+
+        cli.log(
+            chalk.green,
+            chalk.bold(
+                `✅ Successfuly changed '${args[1]}' to '${value}' for user ${user.first_name + " " + user.last_name
+                } id: ${user.id}\n`
+            )
+        );
     } else {
-        cli.log(chalk.red, chalk.bold(`⚠️  Sorry, you can not change '${key}' of any user \n\n`))
+        cli.log(
+            chalk.red,
+            chalk.bold(`⚠️  Sorry, you can not change '${args[1]}' of any user \n`)
+        );
     }
 
     list();
 }
-
 
 function list() {
     cli.printUsersList(User.getUsers());
@@ -99,27 +127,87 @@ function list() {
 
 function read(args: cli.ArgumentsList) {
     const userId = args[0];
-    console.log("printing user", userId)
-    console.log(User.getUserById(args[0]))
+    console.log("printing user", userId);
+    console.log(User.getUserById(args[0]));
 }
 
 function deleteUser(args: cli.ArgumentsList) {
     const userId = args[0];
-    console.log("deleting user", userId)
+    const user = User.getUserById(userId);
+
+    if (!user) {
+        cli.log(
+            chalk.red,
+            chalk.bold(
+                `🗑️ ⚠️  Sorry, couldn't find user with id '${userId}' to delete \n\n`
+            )
+        );
+        list();
+        exit();
+    }
+
+    if (user.soft_deleted) {
+        cli.log(
+            chalk.green,
+            chalk.bold(`🗑️ ✅ User id '${userId}' is already deleted`)
+        );
+        list();
+        exit();
+    }
+
     User.deleteUserById(userId);
     User.saveUsers();
+
+    cli.log(
+        chalk.green,
+        chalk.bold(
+            `🗑️ ✅ Successfuly deleted ${user.first_name + " " + user.last_name} id: ${user.id
+            }\n\n`
+        )
+    );
 }
 
-function restoreUser(args: cli.ArgumentsList) {
+function restore(args: cli.ArgumentsList) {
     const userId = args[0];
+    const user = User.getUserById(userId);
 
+    if (!user) {
+        cli.log(
+            chalk.red,
+            chalk.bold(
+                `♻️ ⚠️  Sorry, couldn't find user with id '${userId}' to restore \n\n`
+            )
+        );
+        list();
+        exit();
+    }
+
+    if (!user.soft_deleted) {
+        cli.log(
+            chalk.green,
+            chalk.bold(`♻️ ✅ User id '${userId}' is already active`)
+        );
+        list();
+        exit();
+    }
+
+    User.restoreUserById(userId);
+    User.saveUsers();
+    cli.log(
+        chalk.green,
+        chalk.bold(
+            `♻️ ✅ Successfuly restored ${user.first_name + " " + user.last_name} id: ${user.id
+            }\n\n`
+        )
+    );
 }
+
 function create(args: cli.ArgumentsList) {
     const user = User.newUser({
         id: User.generateId(),
         first_name: args[0],
         last_name: args[1],
-        phonenumber: args[2]
+        phonenumber: args[2],
     } as User.User);
     User.addUser(user);
     User.saveUsers();
@@ -139,13 +227,12 @@ function main() {
 
     cli.handleCLICommand(
         (cmd, args) => {
-            cmd.callback(args)
+            cmd.callback(args);
         },
         (rawCmd, args) => {
-            cli.printCommandInvalid(rawCmd)
+            cli.printCommandInvalid(rawCmd);
         }
     );
-
 }
 
 main();
@@ -164,17 +251,17 @@ function DEBUG_generateUsers() {
         id: "1",
         first_name: "gilad",
         last_name: "pinker",
-        phonenumber: "054531"
+        phonenumber: "054531",
     } as User.User);
 
     const user2 = User.newUser({
         id: "2",
         first_name: "gilad",
         last_name: "pinker",
-        phonenumber: "054531"
+        phonenumber: "054531",
     } as User.User);
 
-    User.addUser(user1)
-    User.addUser(user2)
-    User.saveUsers()
+    User.addUser(user1);
+    User.addUser(user2);
+    User.saveUsers();
 }
